Update vorob_optim_parallel
There was a problem with the pbivnorm function which doesn't handle +-Inf well in inputs, so I replaced the infinite values of the arg1, arg2 and arg3 terms by +-1000. With this modification, there is no more NaN at the output (this did not happen often in small dimension but in large dimension it became quite annoying).